refactor: DB utils split
