+++ udata_hydra/data_formats/csv_like/to_db.py
@@ -10,11 +10,11 @@ from udata_hydra.analysis import helpers
 from udata_hydra.analysis.tables_index import insert_tables_index_entry
 from udata_hydra.conversion.schema import compute_create_table_query
 from udata_hydra.data_formats import CsvLike
-from udata_hydra.db import compute_insert_query, db_col_name
 from udata_hydra.db.check import Check
 from udata_hydra.db.resource import Resource
 from udata_hydra.utils import ParseException
 from udata_hydra.utils.casting import iter_tabular_rows
+from udata_hydra.utils.db import compute_insert_query, db_col_name
 
 log = logging.getLogger("udata-hydra")
 if TYPE_CHECKING:
