+++ udata_hydra/data_formats/table/to_geojson.py
@@ -7,8 +7,8 @@ from udata_hydra.data_formats.csv_like.to_geojson import (
     DEFAULT_GEOJSON_FILENAME,
     _detect_geo_columns,
 )
-from udata_hydra.db import db_col_name
 from udata_hydra.utils import storage_path
+from udata_hydra.utils.db import db_col_name
 
 if TYPE_CHECKING:
     from udata_hydra.data_formats import Geojson
