+++ udata_hydra/data_formats/table/to_parquet.py
@@ -6,8 +6,8 @@ import pyarrow.parquet as pq
 
 from udata_hydra.conversion.schema import PYTHON_TYPE_TO_PA
 from udata_hydra.data_formats import Table
-from udata_hydra.db import db_col_name
 from udata_hydra.utils import storage_path
+from udata_hydra.utils.db import db_col_name
 
 if TYPE_CHECKING:
     from udata_hydra.data_formats import Parquet
